requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
python-dotenv
google-adk
google-generativeai
reportlab
markdown2